BOUNTY HUNTERS/TASK_14_SLACK_BOT_API/requirements.txt
spacy==3.1.4
slackclient==2.9.3
python-dotenv==0.19.1
